move regression tests to regression.rs
